requirements.txt
flask
python-telegram-bot==13.13
python-dotenv
